Replace the lineinfo segment in Powerline with a new sli segment.
This new segment treats the symbol as it's own segment so that it can be
highlighted separately than the line number.  The reason for this is
because the Hack typeface already has the Powerline symbols, but only in
the regular typeface.  Attempting to show a bolded version under MacVim
would result in an awful replacement character versus the fallback
to regular behavior you would get under Linux.